chore: Move mocks from pkg/internal/mock to /pkg/mock folder
Currently mocks are present in 2 localtions. This PR moves remaining mocks to /pkg/mock folder
