+++ pkg/controller/rest/messaging/operation_test.go
@@ -26,10 +26,10 @@ import (
 	svchttp "github.com/hyperledger/aries-framework-go/pkg/didcomm/messaging/service/http"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/did"
 	"github.com/hyperledger/aries-framework-go/pkg/framework/aries/api/vdri"
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/msghandler"
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/protocol"
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/protocol/generic"
-	mocksvc "github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/service"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/msghandler"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/protocol"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/protocol/generic"
+	mocksvc "github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/service"
 	mockdiddoc "github.com/hyperledger/aries-framework-go/pkg/mock/diddoc"
 	mockkms "github.com/hyperledger/aries-framework-go/pkg/mock/kms/legacykms"
 	"github.com/hyperledger/aries-framework-go/pkg/mock/storage"
