+++ pkg/controller/rest/verifiable/operation_test.go
@@ -28,9 +28,9 @@ import (
 	"github.com/hyperledger/aries-framework-go/pkg/doc/did"
 	verifiableapi "github.com/hyperledger/aries-framework-go/pkg/doc/verifiable"
 	"github.com/hyperledger/aries-framework-go/pkg/framework/aries/api/vdri"
-	mockprovider "github.com/hyperledger/aries-framework-go/pkg/internal/mock/provider"
 	cryptomock "github.com/hyperledger/aries-framework-go/pkg/mock/crypto"
 	kmsmock "github.com/hyperledger/aries-framework-go/pkg/mock/kms"
+	mockprovider "github.com/hyperledger/aries-framework-go/pkg/mock/provider"
 	mockstore "github.com/hyperledger/aries-framework-go/pkg/mock/storage"
 	mockvdri "github.com/hyperledger/aries-framework-go/pkg/mock/vdri"
 )
