+++ pkg/didcomm/messaging/msghandler/example_test.go
@@ -12,7 +12,7 @@ package msghandler
 import (
 	"fmt"
 
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/protocol/generic"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/protocol/generic"
 )
 
 func ExampleRegistrar_Register() {
