+++ pkg/didcomm/messaging/msghandler/msghandler_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/hyperledger/aries-framework-go/pkg/didcomm/dispatcher"
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/protocol/generic"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/protocol/generic"
 )
 
 func TestRegistrar_Register(t *testing.T) {
