+++ pkg/framework/context/context_test.go
@@ -21,16 +21,16 @@ import (
 	"github.com/hyperledger/aries-framework-go/pkg/didcomm/protocol/didexchange"
 	serviceMocks "github.com/hyperledger/aries-framework-go/pkg/internal/gomocks/didcomm/common/service"
 	verifiableStoreMocks "github.com/hyperledger/aries-framework-go/pkg/internal/gomocks/store/verifiable"
-	mockdidcomm "github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm"
-	mockdispatcher "github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/dispatcher"
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/msghandler"
-	mockpackager "github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/packager"
-	mockdidexchange "github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/protocol/didexchange"
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/protocol/generic"
-	mocklock "github.com/hyperledger/aries-framework-go/pkg/internal/mock/secretlock"
 	mockcrypto "github.com/hyperledger/aries-framework-go/pkg/mock/crypto"
+	mockdidcomm "github.com/hyperledger/aries-framework-go/pkg/mock/didcomm"
+	mockdispatcher "github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/dispatcher"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/msghandler"
+	mockpackager "github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/packager"
+	mockdidexchange "github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/protocol/didexchange"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/protocol/generic"
 	mockkms "github.com/hyperledger/aries-framework-go/pkg/mock/kms"
 	mocklegacykms "github.com/hyperledger/aries-framework-go/pkg/mock/kms/legacykms"
+	mocklock "github.com/hyperledger/aries-framework-go/pkg/mock/secretlock"
 	"github.com/hyperledger/aries-framework-go/pkg/mock/storage"
 	mockvdri "github.com/hyperledger/aries-framework-go/pkg/mock/vdri"
 )
