+++ pkg/kms/localkms/internal/keywrapper/kms_aead_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/google/tink/go/tink"
 	"github.com/stretchr/testify/require"
 
-	"github.com/hyperledger/aries-framework-go/pkg/internal/mock/secretlock"
+	"github.com/hyperledger/aries-framework-go/pkg/mock/secretlock"
 )
 
 func TestLocalKMS_New_AEAD(t *testing.T) {
