+++ pkg/mock/didcomm/protocol/mock_protocol.go
@@ -11,9 +11,9 @@ import (
 	"github.com/hyperledger/aries-framework-go/pkg/didcomm/dispatcher"
 	"github.com/hyperledger/aries-framework-go/pkg/didcomm/transport"
 	vdriapi "github.com/hyperledger/aries-framework-go/pkg/framework/aries/api/vdri"
-	mockdispatcher "github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/dispatcher"
-	mockservice "github.com/hyperledger/aries-framework-go/pkg/internal/mock/didcomm/service"
 	"github.com/hyperledger/aries-framework-go/pkg/kms/legacykms"
+	mockdispatcher "github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/dispatcher"
+	mockservice "github.com/hyperledger/aries-framework-go/pkg/mock/didcomm/service"
 	mockkms "github.com/hyperledger/aries-framework-go/pkg/mock/kms/legacykms"
 	mockstore "github.com/hyperledger/aries-framework-go/pkg/mock/storage"
 	mockvdri "github.com/hyperledger/aries-framework-go/pkg/mock/vdri"
